refactor: extract TrendBadge to its own component file
